made few changes
