--- app/components/Sidebar/page.js
@@ -1,44 +1,0 @@
-import React from 'react';
-
-const Sidebar = () => {
-  return (
-    <div className="sidebar">
-      
-      <h3>Joemon Mathew</h3>
-      <ul>
-        <li>
-          
-            <a>About</a>
-         
-        </li>
-        <li>
-     
-             <a>Work</a>
-
-       </li>
-        <li> 
-          
-            <a>Skills</a>
-         
-        </li>
-        <li>
-     
-            <a>Resume</a>
-      
-        </li>
-        <li>
-     
-            <a>Blogs</a>
-
-      </li>
-      <li>
-     
-             <a>Contact</a>
-
-     </li>
-      </ul>
-    </div>
-  );
-};
-
-export default Sidebar;
