--- app/globals.css
@@ -1,46 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body{
-  background-color: black;
-}
-ul{
-  list-style: none;
-}
-.main-content{
-  color:white;
-  margin-left: 25%; 
-  padding: 20px;
-}
-
-.sidebar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 20%;
-  background-color: #333; 
-  color: #4CAF50;
-  padding: 20px;
-  box-sizing: border-box;
- 
-}
-.sidebar h3 {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin-bottom: 10px;
-  box-sizing: border-box;
-}
-
-.sidebar a {
-  color: #4CAF50;
-  text-decoration: none;
-  display: block;
-  margin-bottom: 10px;
-  box-sizing: border-box;
-  
-}
-.sidebar a:hover {
-  color: black; 
-}
