+++ components/MainContent.js
@@ -2,11 +2,11 @@ import React from 'react';
 
 const MainContent = () => {
   return (
-    <div className="main-content"> 
+    <div className='flex flex-col m-10 items-center space-y-4'> 
       <h1>Hello, I'm Joemon Mathew</h1>
       <h3>Frontend Developer </h3>
       <p>
-        Hey there... I will add my bio here 
+        Hey there... I will add my bio here! 
       </p>
     </div>
   );
