+++ pages/index.js
@@ -1,0 +1,17 @@
+import React from "react";
+import MainContent from '../components/MainContent'
+import Sidebar from "../components/Sidebar";
+import Head from "next/head";
+export default function Home() {
+  return (
+   <div className="flex">
+      <Head>
+        <title> Joemon | Front-End Developer</title>
+      </Head>
+      <Sidebar />
+    <MainContent />
+
+     
+   </div>
+  )
+}
