+++ styles/globals.css
@@ -1,0 +1,30 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+
+@layer base{
+  body {
+    @apply bg-black text-white tracking-wide
+  }
+  
+  h1, h2, h3, h4, h5, h6 {
+    @apply font-bold
+  }
+
+  h1{
+    @apply text-4xl sm:text-5xl md:text-6xl
+  }
+
+  h2{
+    @apply text-3xl sm:text-4xl
+  }
+
+  h3{
+    @apply text-2xl sm:text-3xl
+  }
+  li{
+    @apply cursor-pointer
+  }
+}
+
